requirements.txt
autopep8==2.3.2
blinker==1.9.0
click==8.3.1
Flask==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
psycopg2-binary==2.9.11
pycodestyle==2.14.0
Werkzeug==3.0.6